% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCF.R
\name{filterMissVCF}
\alias{filterMissVCF}
\title{Filter a Colony VCF df for sample missingness}
\usage{
filterMissVCF(vcf.df, perc.missing = 50)
}
\arguments{
\item{vcf.df}{modified VCF dataframe from processVCF() function}

\item{perc.missing}{the allowable \% (0-100) of loci for which a sample can be missing data}
}
\value{
A VCF dataframe with some samples removed
}
\description{
This function takes the input from the processVCF() function and filters for
sample missingness. So individuals that are missing more than perc.missing
are dropped from the output.
}
