% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCF.R
\name{processVCF}
\alias{processVCF}
\title{Load and Process a VCF}
\usage{
processVCF(vcf.file)
}
\arguments{
\item{vcf.file}{Path to the input VCF file}
}
\value{
A dataframe of the modified VCF genotype
}
\description{
This function loads in a VCf using the vcfR::read.vcfR. The vcf file is then
processed to match the Colony input format. This function returns an updated
VCF genotype dataframe with modified genotype notation. This dataframe is the
required input for the main function "writeColonyFile()".
}
