% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCF.R
\name{shortNames}
\alias{shortNames}
\title{Reduce the length of column and row names}
\usage{
shortNames(vcf.df, num.row.char = 20, num.col.char = 20)
}
\arguments{
\item{vcf.df}{any dataframe}

\item{num.row.char}{number of desired characters in row names}

\item{num.col.char}{number of desired characters in column names}
}
\value{
A VCF dataframe with shortened column and row names
}
\description{
This function cuts the column and row names of a dataframe.
}
