% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeColonyFile.R
\name{writeColonyFile}
\alias{writeColonyFile}
\title{Write a Colony Input Data File}
\usage{
writeColonyFile(
  colony.run.name = "colony_run1",
  offspring.id = offspring.id,
  father.id = father.id,
  mother.id = mother.id,
  vcf.input.df = vcf.input.df,
  out.path = "./",
  out.file.suffix = "colonyInputFile.txt",
  num.char = 20,
  seed = 1234,
  updateAF = 0,
  MD = 2,
  InBD = 0,
  ploid = 0,
  PM = 0,
  clone = 0,
  SibSizeScale = 0,
  SibSizePrior = 0,
  popAF = 0,
  n.runs = 1,
  l.run = 1,
  monitor = 0,
  monInterval = 10000,
  AnalysisType = 1,
  Precision = 3,
  MarkerTypes = "0@",
  AlleleDropRate = "0.02@",
  OtherErrorRate = "0.02@",
  ProbParentIncl = 0.01
)
}
\arguments{
\item{colony.run.name}{Name for this colony run (included in the filename)}

\item{offspring.id}{List of all potential offspring sample names}

\item{father.id}{List of all potential father sample names}

\item{mother.id}{List of all potential mother sample names}

\item{vcf.input.df}{Modified genotype dataframe generated from the processVCF function}

\item{out.path}{The full path to the desired output folder, if it does not exist it will be created}

\item{out.file.suffix}{The file ending that will be appended to the colony.run.name}

\item{num.char}{The desired length to cut rownames and colnames (must be <= 20 for Colony)}

\item{seed}{<- 1234 The random seed for Colony}

\item{updateAF}{0/1=Not updating/updating allele frequency}

\item{MD}{2/1=Dioecious/Monoecious species}

\item{InBD}{0/1=No inbreeding/inbreeding}

\item{ploid}{0/1=Diploid species/HaploDiploid species}

\item{PM}{0/1=Polygamy/Monogamy for males & females}

\item{clone}{0/1=Clone inference =No/Yes}

\item{SibSizeScale}{0/1=Full sibship size scaling =No/Yes}

\item{SibSizePrior}{0, 1, 2, 3 = No, weak, medium, strong sibship size prior; mean paternal & maternal sibship size}

\item{popAF}{0/1=Unknown/Known population allele frequency}

\item{n.runs}{Number of replicate Colony runs}

\item{l.run}{Length of Colony run. Give a value of 1, 2, 3, 4 to indicate short, medium, long, very long run.}

\item{monitor}{0/1=Monitor method by Iterate. Give a value of either 0 or 1 to indicate monitoring the intermediate results by iterate number or running time. Always choose value 0 for run without Windows GUI.}

\item{monInterval}{Monitor interval in Iterate}

\item{AnalysisType}{Analysis 0 (Pairwise-Likelihood Score), 1 (Full Likelihood), 2 (combined Pairwise-Likelihood Score and Full Likelihood)}

\item{Precision}{1/2/3=low/medium/high Precision for Full likelihood}

\item{MarkerTypes}{Marker Type: 0 co-dominant / 1 dominant, use "0@" or "1@" to repeat the option for all loci}

\item{AlleleDropRate}{Allelic Drop-out rate, use "x@" to repeat the option for all loci}

\item{OtherErrorRate}{Other Error rate, use "x@" to repeat the option for all loci}

\item{ProbParentIncl}{Probability parent included}
}
\value{
Writes a Colony input data file into the specified directory
}
\description{
This function writes a data file for use in the software Colony (https://www.zsl.org/science/software/colony).
This function requires a modified vcf genotype dataframe that can be generated using the processVCF function.
Note: Colony does not accept row or column names longer than 20 characters. This function includes "num.char"
argument to cut these names.
}
